+++ docs/demo/time-picker/section-base.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import TimePicker from '../../../components/date-picker/TimePicker'
 const prefix = 'section-base'
-const rightOptions = ['基础', '禁用']
+const rightOptions = ['基础', '步长设置', '禁用']
 const desc = '选择时间点，可与日期搭配使用，也可用于展示当天时间'
 const code = [
   {
@@ -36,6 +36,23 @@ const code = [
       }
     }`,
     opt: ['基础']
+  },
+  {
+    code: `import React from 'react'
+    import TimePicker from '@hi-ui/hiui/es/date-picker/TimePicker'\n
+    class Demo extends React.Component {
+      render() {
+        return (
+          <TimePicker
+            hourStep={5}
+            minuteStep={6}
+            secondStep={10}
+            onChange={(date, dateString) => console.log(date, dateString)}
+          />
+        )
+      }
+    }`,
+    opt: ['步长设置']
   },
   {
     code: `import React from 'react'